requirements.txt
streamlit
langchain
langchain-openai
langgraph
pandas
python-dotenv
matplotlib
requests